% Auto-generated: do not edit by hand
\name{''DateRangePicker}

\alias{''DateRangePicker}

\title{DateRangePicker component}

\description{
DateRangePicker allows a user to set a date range.
}

\usage{
''DateRangePicker(id=NULL, ariaLabels=NULL, calendarFocus=NULL,
className=NULL, classNames=NULL, color=NULL, date=NULL,
dateDisplayFormat=NULL, dayDisplayFormat=NULL,
direction=NULL, disabledDates=NULL, disabledDay=NULL,
displayMode=NULL, dragSelectionEnabled=NULL,
editableDateInputs=NULL, endDatePlaceholder=NULL,
fixedHeight=NULL, focusedRange=NULL, footerContent=NULL,
headerContent=NULL, initialFocusedRange=NULL,
inputRanges=NULL, locale=NULL, maxDate=NULL, minDate=NULL,
monthDisplayFormat=NULL, months=NULL,
navigatorRenderer=NULL, onChange=NULL, onPreviewChange=NULL,
onRangeFocusChange=NULL, onShownDateChange=NULL,
preventSnapRefocus=NULL, preview=NULL, rangeColors=NULL,
ranges=NULL, renderStaticRangeLabel=NULL, scroll=NULL,
showDateDisplay=NULL, showMonthAndYearPickers=NULL,
showMonthArrow=NULL, showPreview=NULL, shownDate=NULL,
startDatePlaceholder=NULL, staticRanges=NULL,
updateRange=NULL, weekStartsOn=NULL,
weekdayDisplayFormat=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{ariaLabels}{}

\item{calendarFocus}{Character. }

\item{className}{Character. }

\item{classNames}{Named list. }

\item{color}{Character. }

\item{date}{Named list. }

\item{dateDisplayFormat}{Character. }

\item{dayDisplayFormat}{Character. }

\item{direction}{A value equal to: 'vertical', 'horizontal'. }

\item{disabledDates}{Unnamed list. }

\item{disabledDay}{}

\item{displayMode}{A value equal to: 'daterange', 'date'. }

\item{dragSelectionEnabled}{Logical. }

\item{editableDateInputs}{Logical. }

\item{endDatePlaceholder}{Character. }

\item{fixedHeight}{Logical. }

\item{focusedRange}{List of numerics. }

\item{footerContent}{Logical | numeric | character | named list | unnamed list. }

\item{headerContent}{Logical | numeric | character | named list | unnamed list. }

\item{initialFocusedRange}{List of numerics. }

\item{inputRanges}{Unnamed list. }

\item{locale}{Named list. }

\item{maxDate}{Named list. }

\item{minDate}{Named list. }

\item{monthDisplayFormat}{Character. }

\item{months}{Numeric. }

\item{navigatorRenderer}{}

\item{onChange}{}

\item{onPreviewChange}{}

\item{onRangeFocusChange}{}

\item{onShownDateChange}{}

\item{preventSnapRefocus}{Logical. }

\item{preview}{Lists containing elements 'startdate', 'enddate', 'color'.
those elements have the following types:
  - startdate (named list; optional)
  - enddate (named list; optional)
  - color (character; optional). }

\item{rangeColors}{List of characters. }

\item{ranges}{List. }

\item{renderStaticRangeLabel}{}

\item{scroll}{Lists containing elements 'enabled', 'monthheight', 'longmonthheight', 'monthwidth', 'calendarwidth', 'calendarheight'.
those elements have the following types:
  - enabled (logical; optional)
  - monthheight (numeric; optional)
  - longmonthheight (numeric; optional)
  - monthwidth (numeric; optional)
  - calendarwidth (numeric; optional)
  - calendarheight (numeric; optional). }

\item{showDateDisplay}{Logical. }

\item{showMonthAndYearPickers}{Logical. }

\item{showMonthArrow}{Logical. }

\item{showPreview}{Logical. }

\item{shownDate}{Named list. }

\item{startDatePlaceholder}{Character. }

\item{staticRanges}{Unnamed list. }

\item{updateRange}{}

\item{weekStartsOn}{Numeric. }

\item{weekdayDisplayFormat}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

